requirements.txt
Django==3.0.8
Pillow==7.2.0
gunicorn==20.0.4
django-heroku==0.3.1
django-tinymce==3.0.2
whitenoise==5.2.0
python-dotenv==0.14.0
